requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[dotenv]
pydantic-settings
python-multipart
sentence-transformers
faiss-cpu
pypdf
docx2txt
google-generativeai
pytest
pytest-cov
